Fixes issue #2238 - Consolidate request unit tests
